+++ lib/multimotor/src/motordrive.h
@@ -12,5 +12,5 @@ public:
     virtual void setSpeed(float speed) = 0;
     virtual bool handleIncoming(uint32_t id, uint8_t* data, uint8_t len, uint32_t now) = 0;
     virtual uint32_t getLastStatusTime() const = 0;
-    virtual uint8_t getLastFaults() const = 0;
+    virtual uint32_t getLastFaults() const = 0;
 };
